app/contactApp/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions==1.9.0
azure-cosmos==4.2.0
azure-cli-core==2.34.1
bson==0.5.10
pymongo==4.0.2
Flask==1.1.2
requests==2.23.0